+++ db/schema/index.ts
@@ -1,4 +1,5 @@
 export * from './alerts';
+export * from './apiKeys';
 export * from './dailyRollups';
 export * from './enums';
 export * from './errors';